Added modal to confirm leaving course and functionality now works
Modal now pops up when you click the "leave course" button to confirm if you want to leave the course. Clicking confirm works and updates the state of the courses you're currently in on your profile. Still having issues with styling that come from z-index in AboutUser.js in the ContentWrapper.